% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{compare_distributions}
\alias{compare_distributions}
\title{Compare Two Distributions with Zoom in to Tails}
\usage{
compare_distributions(dist1, dist2, zoom_tail = "left", ...)
}
\arguments{
\item{dist1}{a dataframe with 2 columns: ret = returns ; type = column type}

\item{dist2}{a dataframe with 2 columns: ret = returns ; type = column type}

\item{zoom_tail}{currently supports "left"}
}
\value{
ggplot object
}
\description{
Compare Two Distributions with Zoom in to Tails
}
\examples{
library(tidyverse)
library(ggplot2)
library(highfrequency)
library(xts)
library(ggforce)

quotes <- highfrequency::sampleQData \%>\% highfrequency::aggregateQuotes(alignBy = "seconds",alignPeriod = 5)\%>\% select(DT,ret = MIDQUOTE)
log_returns <- as.xts(quotes) \%>\% highfrequency::makeReturns()

normal_dist_logret   <- rnorm(length(log_returns),
                              mean= mean(log_returns),
                              sd = sd(log_returns))

dist1 <- log_returns \%>\% as.data.frame() \%>\% remove_rownames() \%>\% mutate(type = "logret")
dist2 <- normal_dist_logret \%>\% data.frame(ret = .) \%>\% mutate(type = "normdist")
compare_distributions(dist1,dist2,zoom_tail = "left") +
  scale_color_manual(name = "Distributions",
                     values = c("logret" = "red",
                                "normdist" = "black"),
                     labels = c("High Frequency Log Returns","Normal Distribution"))
}
\keyword{tbc}
